делает создание карты экспортируемой функцией
