+++ js/map.js
@@ -1,30 +1,11 @@
-import { turnFormOn, turnFormOff } from './form.js';
-import { turnFiltersOn, turnFiltersOff } from './filters.js';
-import { createAds } from './data.js';
+import { turnFormOn } from './form.js';
+import { turnFiltersOn } from './filters.js';
 import { createPopups } from './popup.js';
 
 const addressField = document.querySelector('#address');
-const similarAds = createAds(10);
 
-turnFiltersOff();
-turnFormOff();
-
-const map = L.map('map-canvas')
-  .on('load', () => {
-    turnFiltersOn();
-    turnFormOn();
-  })
-  .setView({
-    lat: 35.68211,
-    lng: 139.75364,
-  }, 12);
-
-L.tileLayer(
-  'https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png',
-  {
-    attribution: '&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors',
-  },
-).addTo(map);
+const map = L.map('map-canvas');
+const markerGroup = L.layerGroup().addTo(map);
 
 const mainPinIcon = L.icon({
   iconUrl: 'img/main-pin.svg',
@@ -32,23 +13,22 @@ const mainPinIcon = L.icon({
   iconAnchor: [26, 52],
 });
 
-const mainMarker = L.marker(
-  {
-    lat: 35.68211,
-    lng: 139.75364,
-  },
-  {
-    draggable: true,
-    icon: mainPinIcon,
-  },
-);
+const createMainMarker = (coordinate) => {
+  const mainMarker = L.marker(
+    coordinate,
+    {
+      draggable: true,
+      icon: mainPinIcon,
+    },
+  );
 
-mainMarker.addTo(map);
+  mainMarker.addTo(map);
 
-mainMarker.on('moveend', (evt) => {
-  const newAddress = evt.target.getLatLng();
-  addressField.value = `${newAddress.lat.toFixed(5)}, ${newAddress.lng.toFixed(5)}`;
-});
+  mainMarker.on('moveend', (evt) => {
+    const newAddress = evt.target.getLatLng();
+    addressField.value = `${newAddress.lat.toFixed(5)}, ${newAddress.lng.toFixed(5)}`;
+  });
+};
 
 const generalIcon = L.icon({
   iconUrl: 'img/pin.svg',
@@ -56,9 +36,7 @@ const generalIcon = L.icon({
   iconAnchor: [20, 40],
 });
 
-const markerGroup = L.layerGroup().addTo(map);
-
-const createMarker = ({author, offer, location}) => {
+const createMarkers = ({author, offer, location}) => {
   const marker = L.marker(
     location,
     {
@@ -70,7 +48,23 @@ const createMarker = ({author, offer, location}) => {
     .bindPopup(createPopups({author, offer}));
 };
 
-similarAds.forEach((ad) => {
-  createMarker(ad);
-});
+const createMap = (coordinate) => {
+  map.on('load', () => {
+    turnFiltersOn();
+    turnFormOn();
+  });
 
+  map.setView(
+    coordinate, 12);
+
+  L.tileLayer(
+    'https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png',
+    {
+      attribution: '&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors',
+    },
+  ).addTo(map);
+
+  createMainMarker(coordinate);
+};
+
+export {createMarkers, createMap};
